backend/requirements.txt
fastapi==0.104.1
pydantic==2.11.7
python-multipart==0.0.6
python-dotenv==1.1.1
squareup==43.0.0.20250618
gunicorn
uvicorn[standard]
aiohttp==3.9.5
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.7.0
multidict==6.6.4
yarl==1.20.1
